mcp-server/requirements.txt
mcp[cli]>=1.10.1
httpx>=0.28.1
python-dotenv>=1.0.0
